Add files via upload
A function called same which accepts 2 arrays. The function should return true if every value in the array has it's corresponding
value squared in the second array. The frequency of the values must be the same, but the index
of the value can be different (this differs from the arrDiff program in this repository).